+++ docs/source/changelog.rst
@@ -30,6 +30,12 @@ VMAT
       drgs.open_image.filter(size=3)  # this would be ignored in analyze()
       drgs.analyze()
 
+Profiles
+^^^^^^^^
+
+* :bdg-warning:`Fixed` The ``center_idx`` and ``field_edge_idx`` methods of :class:`~pylinac.core.profile.HillProfilePhysical` were
+  not correctly calculated if y-values were being passed in directly. I.e. ``HillProfilePhysical(x=..., y=..., dpmm=None)``. This has been fixed.
+
 v 3.37.0
 --------